{"schema":"decodex/commit/1","summary":"Extract worker live sampling owner","authority":"manual"}
